Attempt at using collections for TIL
Ultimately, this probably won't work. Setting up separate categories for
posts will be a pain, and using collections means I will have to use the
yaml front matter and type in the date there as well. Might not seem
like much work, but I'm trying to make this as frictionless as possible
to get posts written.